mid way through fixing the frames/props.go logic.
my guess is that the reason for the random panics when reading 'ready'
messages is due to the possibility of delimiter chars 'randomly'
appearing within the uuid's. i.e. because the bytes in uuid's are pseudorandom, it's possible for a ':', or '\n' character to appear in one which breaks the logic that reads message frame properties.