passing uuid's as in their raw binary form is proving to be a problem.
i've created a little utility for running checks against 'n' generated
UUIDs to see how often character (byte) sequences might appear.

so far it looks as though double byte sequences such as [0 0] or [13 10]
appear between 150-200 times for every million generated UUIDs.
although that's a very small percentage (~0.015%), it's obviously not
good enough.

the reason for sending UUIDs in their raw binary form was one of
economy (16 bytes as opposed to 36 when formatted as a hyphenated string), but
adding more bytes to the delimiter used between message properties not
only seems wrong, but will start to undermine this decision.

one way to get around this would be simply to qualify (enclose) property
values that shouldn't be parsed e.g. in quotes, braces whatever.