props_test & bytes_test tests all pass, BUT TestPropsToBytes is broken.
what i hadn't considered when writing the TestPropsToBytes was that
maps are unordered and the order in which they're retrieved is
non-deterministic. this means we can't know the sequence of bytes
returned by PropsToBytes which in turn means that TestPropsToBytes passes sometimes
and fails others. a knee-jerk reaction would be to sort the map entries
before putting them into the byte slice, but this would only be to
fix the test and adds otherwise unnecessary complexity. i'm going to
leave things as they are for the time being as (although you shouldn't
ever leave failing tests) they were never a priority of this project.
i'll create an issue and return to it later if/when it makes sense.