still working on the log/trace client/collector example.
i'm having problems synchronizing between the different moving parts and go channels. i think this is mainly because i designed the pub-sub api based on the more common architecture where the pub is a stable component in the architecture and the sub nodes are dynamic. in the logging example the roles have been reversed i.e. multiple dynamic pub endpoints, and a stable subscriber (the collector).